+++ hash2group/Cargo.toml
@@ -28,3 +28,15 @@ p256 = { workspace = true, features = ["arithmetic", "expose-field", "hash2curve
 hex-literal = { workspace = true }
 sha2 = { workspace = true }
 sha3 = { workspace = true }
+
+[[test]]
+name = "rfc9380_expand"
+required-features = ["rfc9380"]
+
+[[test]]
+name = "rfc9380_hash2curve_p256"
+required-features = ["rfc9380", "p256"]
+
+[[test]]
+name = "rfc9380_hash2curve_k256"
+required-features = ["rfc9380", "k256"]
